worker-bee/requirements.txt
kafka-python==2.1.0
docker==6.1.3
python-dotenv==1.0.0
